Code refactor on PetChange View & small changes
